Create an object for Cell
This way its much easier to keep track of the mines and cells.

Also if i wanted to check in the end of the game if the user has won,
i couldn't with just a boolean. Because its necessary to count the number of
flagged cells which are actually bombs. Without this object, i have to count
that from the DOM!